refactor to add a decode module
