+++ eagleproject/notify/triggers/governor_proposals.py
@@ -3,8 +3,8 @@ from eth_hash.auto import keccak
 from eth_abi import decode_single
 from eth_utils import encode_hex, decode_hex
 from django.db.models import Q
-from core.common import decode_calls
 from core.blockchain.addresses import CONTRACT_ADDR_TO_NAME
+from core.blockchain.decode import decode_calls
 from notify.events import event_high
